% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manual.R
\name{pytha}
\alias{pytha}
\title{Find the Hypotenuse of a triangle}
\usage{
pytha(x, y)
}
\arguments{
\item{x}{numeric variable}

\item{y}{numeric variable}
}
\description{
Find the Hypotenuse of a triangle
}
